% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{diceProb}
\alias{diceProb}
\title{Get dice probabilities}
\usage{
diceProb(dice, explodeDepth = 4)
}
\arguments{
\item{dice}{character, a valid dice syntax}

\item{explodeDepth}{if dice syntax include explosion, how many explosions should be calculated}
}
\description{
Get dice probabilities
}
