% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{diceStats}
\alias{diceStats}
\title{Get dice probabilities by rolling}
\usage{
diceStats(dice, n = 1000)
}
\arguments{
\item{dice}{character, a valid dice syntax}

\item{n}{how many times dice should be rolled}
}
\description{
Get dice probabilities by rolling
}
