% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{roll}
\alias{roll}
\title{Roll a dice}
\usage{
roll(dice, critMark = TRUE, vocal = TRUE, returnRolls = FALSE)
}
\arguments{
\item{dice}{character, a valid dice syntax}

\item{vocal}{Should it print individual rolls}

\item{returnRolls}{Logical. If true a list will be returned that includes rolled and dropped dice as well as the sum of accepted dice}
}
\description{
Rolls the dice described as a string
}
