% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{swarm}
\alias{swarm}
\title{Simulate swarm attack}
\usage{
swarm(AC, count, damageDice, attackBonus = 0, damageBonus = 0,
  advantage = c("N", "D", "A"))
}
\arguments{
\item{AC}{AC of the opponent}

\item{count}{number of individuals in the swarm}

\item{damageDice}{damage dice to roll}

\item{attackBonus}{bonus to hit}

\item{damageBonus}{bonus to damage}

\item{advantage}{N = normal, A = with advantage, D = with disadvantage.}
}
\description{
simulates attack by a swarm to an opponent of known AC
}
